Fix pulp app
* Move pulp urls out of galaxy's main application to avoid
  using URL namespace prefix.
* Fix dynaconf not loaded at django startup.
