+++ galaxy/api/urls.py
@@ -254,14 +254,6 @@ collection_urls = [
     url(r'^$', views.UploadCollectionView.as_view()),
 ]
 
-pulp_urls = [
-    url(r'^tasks/$',
-        views.TaskViewSet.as_view({'get': 'list'})),
-    url(r'^tasks/(?P<pk>[^/.]+)/$',
-        views.TaskViewSet.as_view({'get': 'retrieve'}),
-        name='tasks-detail'),
-]
-
 v1_urls = [
     url(r'^$', views.ApiV1RootView.as_view(), name='api_v1_root_view'),
     url(r'^account/', include(account_urls)),
@@ -289,8 +281,6 @@ v1_urls = [
     url(r'^content_blocks/', include(content_block_urls)),
     url(r'^content_types/', include(content_type_urls)),
     url(r'^community_surveys/', include(community_survey_urls)),
-    # NOTE(cutwater): This is temporary location for pulp task URLs
-    url(r'^pulp/', include(pulp_urls)),
 ]
 
 me_urls = [
